[VSA] Fix int32 address overflow in build_l2_aware_lim at long contexts
The composite-key Triton kernel inside build_l2_aware_lim_vsa_qk_fp8_pv_fp4
computed task addresses as

    fkv = tl.load(q2k_idx_ptr + offs * idx_stride0, ...)

with both `offs` and `idx_stride0` typed as int32.  Under the canonical
(n_tasks, max_kv) layout this multiplies up to (n_tasks - 1) * max_kv,
which silently wraps once that product exceeds INT32_MAX (2,147,483,647).
After the wrap the kernel reads from a hugely-negative offset off the
base pointer and trips HSA_STATUS_ERROR_MEMORY_APERTURE_VIOLATION,
surfaced asynchronously as "HIP error: an illegal memory access was
encountered".

The overflow boundary is num_q_blks^2 > 2^31 -- harmless for the test
matrix (BH=4, num_q_blks <= 2669 => product < 30M, ~75x under int32) but
hit hard by Seedance's 4K video path:

    T          = 5,942,400 tokens
    num_q_blks = max_kv = 46,425
    product    = 46,424 * 46,425 = 2,155,234,200  (INT32_MAX + 7.75M)

Fix: declare idx_stride0 as tl.int64 and explicitly promote offs to
int64 before the multiply.  The kernel is bandwidth-bound (two loads,
two stores, a handful of SALU ops) so the extra sign-extend is
unmeasurable.

Verified on gfx950 (MI355X):

  - Seedance scale repro (num_q_blks=46,425, q2k_idx ~8 GB):
    build_l2_aware_lim no longer aperture-violates; returns a valid
    int32 (46_425,) permutation.
  - composite_key_kernel timing constant 13-15 us across num_q_blks
    in {877, 2_669, 8_000, 20_000, 46_425} (bandwidth-bound, int64
    promote invisible).
  - End-to-end op_tests regression (BH=4, T=341,632, 10_676 tiles):
    15.33 ms / call, cos=0.983048, cos(lse)=1.000 -- bit-equal to
    pre-patch baseline.
  - /home/block_sparse.py prod data (1, 4, 112256, 128): 0 NaN / 0 Inf,
    abs_mean=0.17009 unchanged.

+++ aiter/ops/vsa_qk_fp8_pv_fp4.py
@@ -76,7 +76,10 @@ def _vsa_qk_fp8_pv_fp4_composite_key_kernel(
     n_tasks: tl.int32,
     max_kv_x2: tl.int32,
     threshold: tl.int32,
-    idx_stride0: tl.int32,
+    idx_stride0: tl.int64,   # int64 to survive long contexts where
+                             #   max_kv * (n_tasks-1) > INT32_MAX
+                             # (e.g. Seedance 5.9M tokens => max_kv = n_tasks = 46_425,
+                             #  product 2.155e9 overflows int32 by 7.8M).
     BLOCK: tl.constexpr,
     FKV_LOG2: tl.constexpr,
     QN_BITS: tl.constexpr,
@@ -85,7 +88,11 @@ def _vsa_qk_fp8_pv_fp4_composite_key_kernel(
     offs = pid * BLOCK + tl.arange(0, BLOCK)
     mask = offs < n_tasks
     qn = tl.load(q2k_num_ptr + offs, mask=mask, other=0)
-    fkv = tl.load(q2k_idx_ptr + offs * idx_stride0, mask=mask, other=0)
+    # Promote offs to int64 BEFORE the multiply so the address arithmetic
+    # cannot wrap at huge num_q_blks; without this Triton would do an
+    # int32 multiply and then sign-extend the wrapped result, jumping to
+    # a wild GPU address (HSA_STATUS_ERROR_MEMORY_APERTURE_VIOLATION).
+    fkv = tl.load(q2k_idx_ptr + offs.to(tl.int64) * idx_stride0, mask=mask, other=0)
     band = fkv >> FKV_LOG2
     qn_neg = (max_kv_x2 - qn) & ((1 << QN_BITS) - 1)
     key = (band << QN_BITS) | qn_neg
